apps/ARVR/Testing Ground/Server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.11
python-multipart==0.0.7
requests==2.31.0
